Correcion de las Claes: Certificacion.java y  ProfesionalFormacion.java. Se agrego la anotacion @Column(name"nombre_campo") a  todos los atributos de las clases.

+++ src/main/java/com/smty/ApiServiciosProfesionales/Repositories/HabilidadRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface HabilidadRepository extends JpaRepository<Habilidad,Long> {
+public interface HabilidadRepository extends JpaRepository<Habilidad,Integer> {
 }
